LayoutTests/fast/security-labels/set-innerhtml-expected.txt

This tests that setting innerHTML propagates labels properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tag.isOn(target.innerHTML) is false
PASS tag.isOn(target.firstElementChild) is false
PASS tag.isOn(target.innerHTML) is true
PASS tag.isOn(target.firstElementChild) is true
PASS successfullyParsed is true

TEST COMPLETE
This is a test
